DeafaultSelector name  changed to BasicSelector and move into commands module. BasicSelector is now instanced and added to GlWin using createCommand() in basicselectorcommand.py. Added if __debug__ when printing messages in BasicSelector and BasicPainter.
