+++ src/modules/painters/a_painterbasic/basicpainter.py
@@ -7,8 +7,8 @@ from PySide2.QtGui import QSurfaceFormat
 from PySide2.QtWidgets import QMessageBox
 from painters import Painter
 from signals import Signals, DragInfo
-from glvertdatasforhaders import VertDataCollectorCoord3fNormal3fColor4f, VertDataCollectorCoord3fColor4f
-from glhelp import GLEntityType, GLHelpFun, GLDataType
+from a_painterbasic.glvertdatasforhaders import VertDataCollectorCoord3fNormal3fColor4f, VertDataCollectorCoord3fColor4f
+from a_painterbasic.glhelp import GLEntityType, GLHelpFun, GLDataType
 from OpenGL import GL
 from PySide2.QtCore import QCoreApplication
 from geometry import Geometry
@@ -56,7 +56,7 @@ class BasicPainter(Painter):
         Signals.get().selectionChanged.connect(self.onSelected)
         self.paintDevice = 0
         # self.selType = SelModes.FULL_FILL_NEWMESH     # Full geometry by addMeshData
-        self.selType = SelModes.FULL_FILL_SHADER      # Full geometry by shader2
+        self.selType = SelModes.FACET_FILL_GLOFFSET     # Full geometry by shader2
         # self.selType = SelModes.FACET_FILL              # Facet by filled triangle with z-fight compensation
         #self.selType = SelModes.FACET_FILL_GLOFFSET   # Facet by filled triangle with glPolygonOffset to avoid z-fight
         # self.selType = SelModes.FACET_WF              # Facet by wireframe
@@ -670,7 +670,8 @@ class BasicPainter(Painter):
         :param geometry: geometry which mesh data is to be converted
         :return:
         """
-        tsAMD = time.perf_counter()
+        if __debug__:
+            tsAMD = time.perf_counter()
         mesh = geometry.mesh
 
         # color data
@@ -745,9 +746,9 @@ class BasicPainter(Painter):
 
         self.setlistdata_f3xyzf3nf4rgba(key, vertex_data, normal_data, color_data)
         self.setVertexCounter(key, n_faces)
-
-        dtAMD = time.perf_counter() - tsAMD
-        print("Add mesh data total:", dtAMD)
+        if __debug__:
+            dtAMD = time.perf_counter() - tsAMD
+            print("Add mesh data total:", dtAMD)
         return
 
     def createVertexData(self, fv_indices_flattened, points):
@@ -899,8 +900,6 @@ class BasicPainter(Painter):
         self.setlistdata_f3xyzf3nf4rgba(key, vertices, normals, colors)
         self.setVertexCounter_byNum(key, n_vertices)
 
-        print("BP")
-
     def addMeshdata4oglmdl_poly(self, key, geometry):
         """
         Converts the mesh data of a polygon mesh to the vertex data necessary for OpenGL
@@ -908,7 +907,8 @@ class BasicPainter(Painter):
         :param geometry: geometry holding the mesh data which is to be converted
         :return:
         """
-        tsAMD = time.perf_counter()
+        if __debug__:
+            tsAMD = time.perf_counter()
         mesh = geometry.mesh
 
         # color data
@@ -941,8 +941,9 @@ class BasicPainter(Painter):
 
         self.addArrays4oglmdl_poly(key, fv_indices_np, ar_points, face_normals_np, cstype, c, ar_face_colors, ar_vertex_colors)
 
-        dtAMD = time.perf_counter() - tsAMD
-        print("Add mesh data total:", dtAMD)
+        if __debug__:
+            dtAMD = time.perf_counter() - tsAMD
+            print("Add mesh data total:", dtAMD)
         return
 
     def addArrays4oglmdl_poly(self, key, fv_indices, points, face_normals, cstype, c, face_colors, vertex_colors):
